feat : 커버의 color를 상수로 파일분리
